+++ src/AppInstallerCommonCore/Public/winget/ManifestValidation.h
@@ -75,6 +75,7 @@ namespace AppInstaller::Manifest
         WINGET_DEFINE_RESOURCE_STRINGID(SchemaHeaderManifestVersionMismatch);
         WINGET_DEFINE_RESOURCE_STRINGID(SchemaHeaderUrlPatternMismatch);
         WINGET_DEFINE_RESOURCE_STRINGID(InvalidPortableFiletype);
+        WINGET_DEFINE_RESOURCE_STRINGID(InvalidFontFiletype);
     }
 
     struct ValidationError
@@ -252,6 +253,13 @@ namespace AppInstaller::Manifest
     };
 
     static const std::unordered_set<std::filesystem::path> s_AllowedPortableFiletypes = { L".exe" };
+    static const std::unordered_set<std::filesystem::path> s_AllowedFontFiletypes = {
+        L".otf",         // OpenType Font
+        L".ttf",         // TrueType Font
+        L".fnt",         // Font
+        L".ttc",         // TrueType Font Collection
+        L".otc",         // OpenType Font Collection
+    };
 
     // fullValidation: bool to set if manifest validation should perform extra validation that is not required for reading a manifest.
     std::vector<ValidationError> ValidateManifest(const Manifest& manifest, bool fullValidation = true);